Change introduction with new focus of paper
Which will be the fact that it's relatively easy to come up with new configurations with an evolutionary algorithm, so we will keep it running for more secure configurations.

This closes #38.

Also closes #18 since this actually a kind of seeded evaluation; every generation seeds the next.

This closes #14 since actually evolution can't go any further...